modulize the heap features into angelheap
