+++ .gdbinit
@@ -1,2 +1,3 @@
 source ~/peda/peda.py
 source ~/Pwngdb/pwngdb.py
+source ~/Pwngdb/angelheap/gdbinit.py
