+++ angelheap/command_wrapper.py
@@ -1,0 +1,70 @@
+import gdb
+import angelheap
+
+from utils import *
+
+angelheap_cmd = None
+
+class AngelHeapCmd(object):
+    commands = []
+    def __init__(self):
+        # list all commands
+        self.commands = [cmd for cmd in dir(self) if callable(getattr(self, cmd)) ]  
+
+    def tracemalloc(self,*arg):
+        """ Trace the malloc and free and detect some error """
+        (option,) = normalize_argv(arg,1)
+        if option == "on":
+            try :
+                angelheap.trace_malloc()
+            except :
+                print("Can't create Breakpoint")
+        else :
+            angelheap.dis_trace_malloc()
+
+    def heapinfo(self):
+        """ Print some information of heap """
+        angelheap.putheapinfo()
+
+    def printfastbin(self):
+        """ Print the fastbin """
+        angelheap.putfastbin()
+
+    def inused(self):
+        """ Print the inuse chunk """
+        angelheap.putinused()
+
+
+class AngelHeapCmdWrapper(gdb.Command):
+    """ angelheap command wrapper """
+    def __init__(self):
+        super(AngelHeapCmdWrapper,self).__init__("angelheap",gdb.COMMAND_USER)
+
+    def invoke(self,args,from_tty):
+        global angelheap_cmd
+        self.dont_repeat()
+        arg = args.split()
+        if len(arg) > 0 :
+            cmd = arg[0]
+
+            if cmd in angelheap_cmd.commands :
+                func = getattr(angelheap_cmd,cmd)
+                func(*arg[1:])
+            else :
+                print("Unknown command")
+        else :
+            print("Unknow command")
+
+        return 
+
+class Alias(gdb.Command):
+    """ angelheap Alias """
+
+    def __init__(self,alias,command):
+        self.command = command
+        super(Alias, self).__init__(alias,gdb.COMMAND_NONE)
+
+    def invoke(self,args,from_tty):
+        self.dont_repeat()
+        gdb.execute("%s %s" % (self.command,args))
+
