+++ angelheap/utils.py
@@ -1,0 +1,26 @@
+
+def to_int(val):
+    """
+    Convert a string to int number
+    """
+    try:
+        return int(str(val), 0)
+    except:
+        return None
+
+def normalize_argv(args, size=0):
+    """
+    Normalize argv to list with predefined length
+    """
+    args = list(args)
+    for (idx, val) in enumerate(args):
+        if to_int(val) is not None:
+            args[idx] = to_int(val)
+        if size and idx == size:
+            return args[:idx]
+
+    if size == 0:
+        return args
+    for i in range(len(args), size):
+        args += [None]
+    return args
